requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.4
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
currency-exchange==1.0.1
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.4
django-environ==0.4.5
django-extensions==3.1.3
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
drf-yasg==1.20.0
duplicity==0.8.12.0
english-words==1.0.3
entrypoints==0.3
environ==1.0
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
futures==3.1.1
googletrans==3.0.0
goslate==1.5.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httplib2==0.14.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.8
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.9
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2-binary==2.8.6
pycairo==1.16.2
pycups==1.9.73
PyDictionary==2.0.1
PyGObject==3.36.0
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rfc3986==1.4.0
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.0.1
sqlparse==0.4.1
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.4.3
wadllib==1.3.3
wikipedia==1.4.0
xkit==0.0.0
